Update v5, v6, PHP8

+++ ImportRelevance.php
@@ -3,8 +3,6 @@ require_once __DIR__ . DIRECTORY_SEPARATOR.'ImportFromFile.php';
 
 class ImportRelevance extends ImportFromFile
 {
-    use AppTrait;
-
     /** @var Question  */
     public $currentModel;
 
@@ -23,7 +21,7 @@ class ImportRelevance extends ImportFromFile
         $this->questionCodeColumn = 'code';
 
         if (empty($this->currentModel)) {
-            $this->addError('currentModel', "Unable to find model for row " . json_encode($attributes));
+            $this->addError('currentModel', "Unable to find model for row " . var_dump($attributes));
             return null;
         }
 
@@ -41,7 +39,7 @@ class ImportRelevance extends ImportFromFile
             return $result;
         }
 
-        $this->addError('currentModel', "Unable to save model for row: " . json_encode($attributes));
+        $this->addError('currentModel', "Unable to save model for row: " . serialize($attributes));
 
         $this->failedModelsCount ++;
         return $result;
@@ -52,13 +50,16 @@ class ImportRelevance extends ImportFromFile
      * @param QuestionGroup $model
      * @return bool|int
      */
-    private function  updateGroup(QuestionGroup  $model) {
-        if($model->validate([$this->relevanceAttribute])) {
-            $criteria = new CDbCriteria();
-            $criteria->addCondition('sid=:sid');
-            $criteria->addCondition('gid=:gid');
-            $criteria->params = [':gid' => $model->gid, ':sid'=> $this->survey->primaryKey];
-            return QuestionGroup::model()->updateAll([$this->relevanceAttribute => $model->{$this->relevanceAttribute}], $criteria);
+    private function  updateGroup($model) {
+        if($model instanceof QuestionGroup) {
+            if($model->validate([$this->relevanceAttribute])) {
+                $criteria = new CDbCriteria();
+                $criteria->addCondition('sid=:sid');
+                $criteria->addCondition('gid=:gid');
+                $criteria->params = [':gid' => $model->gid, ':sid'=> $this->survey->primaryKey];
+                return QuestionGroup::model()->updateAll([$this->relevanceAttribute => $model->{$this->relevanceAttribute}], $criteria);
+            }
+            // TODO error?
         }
         return false;
     }
@@ -67,16 +68,19 @@ class ImportRelevance extends ImportFromFile
      * @param Question $model
      * @return bool|int
      */
-    private function  updateQuestion(Question $model) {
-        if($model->validate([$this->relevanceAttribute])) {
-            $criteria = new CDbCriteria();
-            $criteria->addCondition('qid=:qid');
-            $criteria->params = [':qid' => $model->qid];
+    private function  updateQuestion($model) {
+        if($model instanceof Question) {
+            if($model->validate([$this->relevanceAttribute])) {
+                $criteria = new CDbCriteria();
+                $criteria->addCondition('qid=:qid');
+                $criteria->params = [':qid' => $model->qid];
 
-            // delete all manually set conditions
-            Condition::model()->deleteAll('qid=:qid', [':qid' => $model->qid]);
+                // delete all manually set conditions
+                Condition::model()->deleteAll('qid=:qid', [':qid' => $model->qid]);
 
-            return Question::model()->updateAll([$this->relevanceAttribute => $model->{$this->relevanceAttribute}], $criteria);
+                return Question::model()->updateAll([$this->relevanceAttribute => $model->{$this->relevanceAttribute}], $criteria);
+            }
+            // TODO error?
         }
         return false;
 
@@ -85,10 +89,8 @@ class ImportRelevance extends ImportFromFile
     /**
      * @param $row
      * @return Question|QuestionGroup|null
-     * @throws Exception
      */
-    private function findModel($row)
-    {
+    private function findModel($row) {
 
         $this->relevanceAttribute = 'relevance';
 
@@ -114,33 +116,17 @@ class ImportRelevance extends ImportFromFile
     /**
      * @param $row
      * @return QuestionGroup|null
-     * @throws Exception
      */
-    protected function findGroup($row)
-    {
+    protected function findGroup($row) {
         $criteria = new CDbCriteria();
+        $criteria->addCondition('language=:language');
         $criteria->params[':language'] = $this->language;
-        $criteria->params[':sid'] = $this->survey->primaryKey;
-        $criteria->params[':name']=$row['group'];
-
-        $criteria->addCondition('t.language=:language');
-        $criteria->addCondition('t.group_name=:name');
-
-        if($this->isV4plusVersion()) {
-            $criteria->addCondition('group.sid=:sid');
-            /** @var QuestionGroupL10n $l10n */
-            $l10n = QuestionGroupL10n::model()
-                ->with('group')
-                ->find($criteria);
-            if($l10n === null) {
-                throw new ErrorException("Unable to find group with name: " . $row['group']);
-            }
-
-            return $l10n->group;
-
-        }
 
         $criteria->addCondition('sid=:sid');
+        $criteria->params[':sid'] = $this->survey->primaryKey;
+
+        $criteria->addCondition('group_name=:name');
+        $criteria->params[':name']=$row['group'];
         return QuestionGroup::model()->find($criteria);
     }
 
@@ -157,9 +143,10 @@ class ImportRelevance extends ImportFromFile
             return null;
         }
 
-
-
         $criteria = new CDbCriteria();
+        $criteria->addCondition('language=:language');
+        $criteria->params[':language'] = $this->language;
+
         $criteria->addCondition('sid=:sid');
         $criteria->params[':sid'] = $this->survey->primaryKey;
 
@@ -168,11 +155,6 @@ class ImportRelevance extends ImportFromFile
         $criteria->params[':parent_qid'] =$parent->qid;
         $criteria->params[':code'] =$row['code'];
         $this->questionCodeColumn = 'code';
-
-        if(!$this->isV4plusVersion()) {
-            $criteria->addCondition('language=:language');
-            $criteria->params[':language'] = $this->language;
-        }
 
         return Question::model()->find($criteria);
     }
@@ -184,21 +166,18 @@ class ImportRelevance extends ImportFromFile
     protected function findQuestion()
     {
         $criteria = new CDbCriteria();
+        $criteria->addCondition('language=:language');
+        $criteria->params[':language'] = $this->language;
+
         $criteria->addCondition('sid=:sid');
         $criteria->params[':sid'] = $this->survey->primaryKey;
+
+
+        $criteria->addCondition('parent_qid=0');
         $criteria->addCondition('title=:code');
         $criteria->params[':code'] = $this->rowAttributes[$this->questionCodeColumn];
-        $criteria->addCondition('parent_qid=0');
-
-
-        if(!$this->isV4plusVersion()) {
-            $criteria->addCondition('language=:language');
-            $criteria->params[':language'] = $this->language;
-        }
-        return Question::model()->find($criteria);
+        $question = Question::model()->find($criteria);
+        return $question;
     }
-
-
-
 
 }
